+++ R/LearnerFcstForecast.R
@@ -19,7 +19,7 @@ LearnerFcstForecast = R6Class(
       super$.train(task)
       pv = self$param_set$get_values(tags = "train")
       parallel_arg = private$.parallel_arg
-      if (!is.null(parallel_arg) && !is.null(pv$num.cores) && is.null(pv[[parallel_arg]])) {
+      if (!is.null(parallel_arg) && test_int(pv$num.cores, lower = 2L) && is.null(pv[[parallel_arg]])) {
         pv[[parallel_arg]] = TRUE
       }
       private$.set_context(private$.fit(task, pv), task)